requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.7.14
charset-normalizer==3.4.2
Eel==0.18.2
et_xmlfile==2.0.0
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.3
idna==3.10
importlib_resources==6.5.2
macholib==1.16.3
mysql-connector-python==9.3.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
tkinterdnd2==0.4.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zope.event==5.1
zope.interface==7.2
